Default to fully expanding copied values
I hit this today - if a struct is too big, the formatted version will be truncated. Setting that limit to `:infinity` seems like a sane default, as this was the reason I went looking for a clipboard library in the first place (that is, not wanting to deal with copying a huge struct from the terminal).

I'm happy to take time to expose the option and/or add tests if you'd prefer it.